+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `postproc` module.
+- `ElementType::is_gradient`
+
 ### Changed
 - All warnings will be printed to stderr now.